+++ src/test/java/com/matdang/seatdang/waiting/repository/query/WaitingQueryRepositoryTest.java
@@ -41,7 +41,7 @@ class WaitingQueryRepositoryTest {
                             .waitingNumber(i)
                             .waitingOrder(i)
                             .storeId(1L)
-                            .customerInfo(new CustomerInfo(i, "010-1111-1111", ((long) (Math.random() * 3 + 1))))
+                            .customerInfo(new CustomerInfo(i, "010-1111-1111", ((int) (Math.random() * 3 + 1))))
                             .waitingStatus(value)
                             .visitedTime(null)
                             .build());
@@ -54,7 +54,7 @@ class WaitingQueryRepositoryTest {
                     .waitingNumber(i)
                     .waitingOrder(i)
                     .storeId(2L)
-                    .customerInfo(new CustomerInfo(i, "010-1111-1111", ((long) (Math.random() * 3 + 1))))
+                    .customerInfo(new CustomerInfo(i, "010-1111-1111", ((int) (Math.random() * 3 + 1))))
                     .waitingStatus(WaitingStatus.WAITING)
                     .visitedTime(null)
                     .build());
@@ -97,7 +97,7 @@ class WaitingQueryRepositoryTest {
                     .waitingNumber(i)
                     .waitingOrder(10 - i)
                     .storeId(3L)
-                    .customerInfo(new CustomerInfo(i, "010-1111-1111", ((long) (Math.random() * 3 + 1))))
+                    .customerInfo(new CustomerInfo(i, "010-1111-1111", ((int) (Math.random() * 3 + 1))))
                     .waitingStatus(WaitingStatus.WAITING)
                     .visitedTime(null)
                     .build());
